unicollab_main-main/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
pydantic-settings
supabase
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
